requirementsRaspberry.txt
aiofiles==23.1.0
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.12.2
boto3==1.28.10
botocore==1.31.10
bs4==0.0.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.4
cryptography==38.0.3
-e git+https://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@github.com/drftrcld/pokemonCardUi.git@75179f42dc13bfc5293368732be0079fb7bdaec8#egg=DatabaseImageMatching&subdirectory=DatabaseImageMatchingLibrary
-e git+https://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@github.com/drftrcld/pokemonCardUi.git@75179f42dc13bfc5293368732be0079fb7bdaec8#egg=demoUI&subdirectory=demoUI
Django==4.2.2
evdev==1.6.0
Flask==2.1.2
frozenlist==1.3.3
future==0.18.3
idna==3.4
image==1.5.33
importlib-metadata==6.7.0
iso8601==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
keyboard==0.13.5
MarkupSafe==2.1.1
multidict==6.0.4
numpy==1.25.0
opencv-python==4.7.0.72
pandas==1.4.2
pbkdf2==1.3
pigpio==1.78
Pillow==9.5.0
pycparser==2.21
pynput==1.7.6
pyOpenSSL==22.1.0
pyserial==3.5
python-dateutil==2.8.2
python-xlib==0.32
pytz==2022.1
PyYAML==6.0.1
requests==2.31.0
rpi-hardware-pwm==0.1.4
RPi.GPIO==0.7.1
s3transfer==0.6.1
-e git+https://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@github.com/drftrcld/pokemonCardUi.git@75179f42dc13bfc5293368732be0079fb7bdaec8#egg=scservosdk&subdirectory=scservosdk
SimulRPi==0.1.0a1
six==1.16.0
sklearn==0.0.post7
soupsieve==2.4.1
sqlparse==0.4.4
typing_extensions==4.6.3
urllib3==1.26.16
Werkzeug==2.1.2
wifi==0.3.8
yarl==1.9.2
zipp==3.15.0
